backend/pyproject.toml
[project]
name = "demo_crew"
version = "0.1.0"
description = "demo_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13.2"
dependencies = [
    "crewai>=0.11.0,<0.12.0",
    "pyyaml>=6.0.1",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.2",
    # Minimal dependencies required for basic functionality
]

[project.optional-dependencies]
tools = [
    "opencv-python>=4.8.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "openai-whisper>=20240930",  # Changed from faster-whisper to openai-whisper
]

[project.scripts]
demo_crew = "demo_crew.main:run"
run_crew = "demo_crew.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
